Add new terrain and raycast class. Also create some utily maths function as bounding box & lerp. Add the bounding box to the geometry class. Minor improvment and refactoring. Also add new tests in the deferred renderer (temp). Fix bugs.

+++ Razor/src/Razor/Cameras/FPSCamera.cpp
@@ -54,7 +54,7 @@ namespace Razor {
 		if (vp != nullptr) 
 		{
 			isViewportHovered = vp->getIsHovered();
-			aspect_ratio = (float)vp->getSize().x / (float)vp->getSize().y;
+			aspect_ratio = vp->getSize().x / vp->getSize().y;
 		}
 
 		if (capture)
